w2/tutorial/visit-tracker/requirements.txt
Flask==2.1.0
google-cloud-datastore==2.15.1
google-auth==2.17.3
requests==2.28.2